src/main/webapp/app/entities/quiz-bowl-user/delete/quiz-bowl-user-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { QuizBowlUserService } from '../service/quiz-bowl-user.service';

import { QuizBowlUserDeleteDialogComponent } from './quiz-bowl-user-delete-dialog.component';

describe('QuizBowlUser Management Delete Component', () => {
  let comp: QuizBowlUserDeleteDialogComponent;
  let fixture: ComponentFixture<QuizBowlUserDeleteDialogComponent>;
  let service: QuizBowlUserService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [QuizBowlUserDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(QuizBowlUserDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(QuizBowlUserDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(QuizBowlUserService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
